# apps/backend/synapse/contracts/synapse-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Backend API",
    "version": "1.0.0",
    "description": "Shared REST contract implemented by Synapse EE and Cerebro."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/v1/councils": {
      "get": {
        "tags": ["Councils"],
        "summary": "List councils",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Council list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilListResponse" } } } }
        }
      },
      "post": {
        "tags": ["Councils"],
        "summary": "Create a council",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilCreate" } } } },
        "responses": {
          "201": { "description": "Council created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilResponse" } } } },
          "202": { "description": "Council accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilResponse" } } } }
        }
      }
    },
    "/v1/councils/{id}": {
      "get": {
        "tags": ["Councils"],
        "summary": "Get a council",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CouncilId" }],
        "responses": {
          "200": { "description": "Council", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/councils/{id}/close": {
      "post": {
        "tags": ["Councils"],
        "summary": "Close a council",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/CouncilId" }],
        "responses": {
          "200": { "description": "Closed council", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CouncilResponse" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "CouncilId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "responses": {
...
      "NotFound": { "description": "Not found" },
...
    },
    "schemas": {
      "Council": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": ["string", "null"] },
          "question": { "type": "string" },
          "status": { "type": "string" },
          "chairman_model": { "type": ["string", "null"] },
          "members": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "verdict": { "type": ["object", "string", "null"], "additionalProperties": true },
          "settings": { "type": "object", "additionalProperties": true }
        },
        "required": ["id", "question", "status"]
      },
      "CouncilCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "question": { "type": "string" },
          "chairman_model": { "type": "string" },
          "members": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "settings": { "type": "object", "additionalProperties": true }
        },
        "required": ["question"]
      },
      "CouncilResponse": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/Council" }
        },
        "required": ["data"]
      },
      "CouncilListResponse": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Council" } }
        },
        "required": ["data"]
      },
...
    }
  }
}